polish SearchService
